Zuul - gerrit connections - handle mandatory/optional parameters
The GerritConnection struct uses 'omitempty' for optional
parameters. When a parameter is omitted the resulting struct after
deserialization does not own the field. To check that the struct own the
field, we use the default value check which is to check if the struct
field own an empty string (we only use string field).

Change-Id: I2f5b5d8ef1e55f525fa7593655e38b906fa04f98