Support for config-locations in SF Spec
When no config-repo location is set and Gerrit enabled
the config repo is created on Gerrit then populated by
a batch job.

When the config-location is set then the config-repo
script (see below) expects to be able to connect as
ssh to the remote location using the "sf-admin-key" secret
private key.

The config-repo script that populate the config repository
adds the resources and gerrit directory trees.
The script is a simple bach script that could be improved
by a rewrite in Python.

Change-Id: I4778d7ba826394fb63a4a8570816d389c36672f3