Enable keystore/truststore and https for keycloak
The truststore is needed to enable communication with keycloak clients (application that delegate
authentication to keycloak). https://www.keycloak.org/server/enabletls
Internal SF clients (like Gerrit) need to be able to communicate with Keycloak through TLS.

The postInit kcadm command validate that keycloak handles https on 8443.

Communication between nginx-ingress and keycloak is still over HTTP and we rely
on nginx to implement the TLS endpoint for external requests.

Change-Id: I7007a78f68bc00196d3a9a336bf897b3c2c83e0a