gerrit - add gsku sidecar container in pod
This sidecar container listens to the MQTT event stream.
When an authentication is done through Keycloak (on Github)
a connection event is sent over MQTT and this service fetchs the
user's SSH key on Keycloak then updates the Gerrit user
account with the SSH key.

This mimics the sf-config behavior.

Change-Id: I7e2186d1d02e29e4adbff1e347b51dcaf97b91bc