Enable an initial config-check/config-update flow
This commit enable the sf-operator to have a minimal
flow for the config repository with config-check job
in check/gate pipelines and a config-update in the post
pipeline.

The config-update uses a long lived token from the "default"
service account in the namespace.

config-update perform a "kubectl api-resources" to validate
API access.

Change-Id: I1cf2807ef2d894776614ba4aab73365b4c4db2ad