generate Zuul tenant config (main.yaml) from the SF Resources
This change implements adds a sidecar container inside the zuul-scheduler
pod. The purpose of teh sidecar container is to generate the Zuul tenant config file
from the resources system and copy the generated main.yaml file
to the zuul-scheduler container via the shared volume.

When all services are deployed we exec:
- on the zuul sidecar: the generate-tenant-config script
- then, on the zuul-scheduler container: the zuul-scheduler
  full-reconfigure

Note that implementation follow the sidecar option of
https://softwarefactory-project.io/r/c/software-factory/sf-operator/+/26700

Depends-on: https://softwarefactory-project.io/r/c/software-factory/dhall-software-factory/+/26715/

Change-Id: I5dfd2c5583e0b22c507791818f20734be0287d75