Add feature to create headless service; add serviceName to Statefulset
This commit adds the serviceName parameter into the function that
creates StatefulSet. This parameter was missing and accroding to
the OpenShift documentation it is required [1]. Without the serviceName,
name resolution will not work corretly inside the cluster.
Also added feature to create the headless service, that will be used by
Zuul Executor service. When the CI job is computed by the Zuul Executor,
the Zuul Web when connects to the finger service to get the console
logs need to know exactly on which executor service the job is computed.
That's why the Web service needs to be able to resolve the pod ip
address, not just a service (communication via service resolution will
connect to different Executor pods if there are more than 1 pod running,
which might not have the current job running).

This commit fixes issue, that Zuul console was not available and the
Zuul Web can not reach zuul-executor-0.

[1] https://docs.openshift.com/container-platform/4.12/rest_api/workloads_apis/statefulset-apps-v1.html#spec

Change-Id: I76b950943ba61b32aaac0994aab545b04518b41b