[Tests] Make the run-ci-tests.sh functional between redeployment
By ensuring we clear '{{ config_path }}' first then we are able
to ensure the gerrit admin key is the right one in .git/config
for the gerrit remote.

W/o that change if the '{{ config_path }} exists then we wrongly
assume that the remote is correct.

This introduce a simplification by just removing the local copy
of config repository.

Change-Id: Ic73c74bec929428aeeffa603727d6d0d22bda4e8