Add default vars for Pod and Container Security Context
OpenShift comes with some PodSecurityStandard (SCC)
https://kubernetes.io/docs/concepts/security/pod-security-standards/

At the current state of the sf-operator (running in dev mode) OpenShift
allow us to run pods/containers bypassing the SCC. However
it reports violation of the default "restricted" SCC.

This change provides default Security Context definition for
pods and containers that satisfied the "restricted" SCC.

We should satified the "restricted" SCC on followup changes.

The experience gained from the running the monocle-operator
in a dedicated namespace / and with a service-user shown that
the SCC violation prevented the operator Pod to run.

It seems also that defining clear Security requirementw
as SecurityContext might be a good practice to start soon
in the development of the sf-operator as it will have an impact
on how we build container images and for further certifications.

Change-Id: I6c8f3d5dcd06fff8c919d8647fcd6041d0a8aca1