Add storage spec for services
Each service has a CRD entry for storage size and type. The size parameter
is of Quantity type (see https://pkg.go.dev/k8s.io/apimachinery/pkg/api/resource#Quantity )

By default, if not explicitly set in the CRD, each size value
is set to 1Gi (but can be set properly for each service), and storage
classname is set to the default value at the root of the spec.

Some services have several storage configs: for example, a deployer
can specify the DB storage size and the log storage size for the
mariaDB service.

Change-Id: I6aaeaa92dda66a8058e1879f530d4773f6da8bc0