Change logserver from statefulset to deployment
The statefulset does not share same volume across the pods. It means,
that when we scale up the logserver, it would create new volume to store
the data [1]. It would not be working on non shared storage backend (NFS,
Ceph etc.).
The deployment creates one volume and it shares same volume across the pods,
so it should be better for logserver deployment. [2][3]

[1] https://kubernetes.io/docs/concepts/workloads/controllers/statefulset/
[2] https://kubernetes.io/docs/concepts/workloads/controllers/deployment/
[3] https://www.baeldung.com/ops/kubernetes-deployment-vs-statefulsets

Change-Id: Ib60cc5c3ef2b7fbc58f02f35ccb3baf4a89b8d85