Remove mandatory 'namespace' param
The operator must be able to watch all namespace when no namespace
is specified.

Generated manifest to run the operator (in a non dev mode) via a
Deployment resources does not expected a fixed namespace and nothing
specific require a hardcoded namespace or the need to specified a
given namespace.

This this change introduces a WATCH_NAMESPACE env var to be set
when we want the operator to watch a single namespace, if not
specified it watches all namespaces.

The '--namespace' is still supported but if not specified we fallback
on the WATCH_NAMESPACE technic.

When running dev/standalone mode by passing the CR path as paramater the
ns is mandatory so we exit with error if not specified with the dev mode.

Change-Id: I42fa1597cf0603d13e101ecb01599982b2403d27