Use OLM bundle to install the operator
This change leverages the "operator bundle run" facility [1] to
install the sf-operator via OLM on the testing node.

"operator-sdk run bundle" command expects an operator-bundle image on
a regular registry so this change also setup a temporary registry.

To summarize the operator run bundle commands performs the following
tasks:

- It creates a OLM Catalog including the operator version described by
  the bundle (CSV)
- It install the Catalog resources on the test node
- It create a OLM Subscription to the sf-operator
- It ensures that the OLM InstallPlan is validated and that the CRD is available

Then a SofwareFactory CR can be reclaimed as usual.

[1]: https://sdk.operatorframework.io/docs/olm-integration/tutorial-bundle/#deploying-an-operator-with-olm

Change-Id: I5756b803a7e63d7706b99c82455fa07f36fea64e