Update the wipe-deployment process since using OLM deployment for CI
The `run-ci-tests.sh` script uses OLM (like for the CI) to deploy the
sf-operator and then reclaim of SF instance. The sf-operator is then
running in the `bundle-catalog-ns` namespace which might need to be
cleaned when a developer will run the operator via `go run` to avoid
two operators watching for the same CRD.

This changes update the wipe-deployment.sh script to re-use the
`clean-installations` role.

Change-Id: I372d601745f983b1c2cd1e8850dfdfaab6c29e8c