Only keep one CR for SoftwareFactory
The process described in the CONTRIBUTING.md where:
- cp config/samples/sf_v1_softwarefactory.yaml my-sf.yaml
- go run ./main.go --namespace sf --cr ./my-sf.yaml
- ./tools/run-ci-tests.sh --tags test_only

failed on the test to ensure the logs on logserver have been
purged. It failed because the loopDelay was not set to 5s
in the config/sample/ file.

This change ensures that in the CI and in the CONTRIBUTING.md
doc we reference the same SoftwareFactory CR file. Thus this change
removes allinone.yaml and replace it by a symlink to the one
in config/sample/.

Change-Id: I053b6ef67db89022be9e7f480817f03ade5071fc