sfconfig: Add the gerrit command
This new command deploy or wipe a Gerrit deployment + belonging
resources into a dedicated namespace named gerrit.

./tools/sfconfig gerrit --deploy
./tools/sfconfig gerrit --wipe

The deployment is similar to the one done by the sf-operator
manager.

This change is transitional and additional changes will follow-up
to fully extract Gerrit from the manager in favor of that new command.

Change-Id: I0f16e8985284dbf2dba76e4e005a8f50404c0c87