Add the "reconciledBy" status field
This change adds a new field status.reconciledBy which aim to
give insight to the external watcher about the which version
of the manager last updated the status field.

This is handy for us in the CI where we need to ensure that data we rely
on are accurate and up to date. Thus having that info is crutial for the upgrade
job to ensure that beyond the fact the manager (sf-operator) is started
that the Reconcile loop at run at least once with the update sf-operator
code.

This change also updates the CI flow:
- To rely on the new 'reconciledBy' field
- By removing the custom resource apply after the upgrade
  as the resource is already applied before the upgrade.

Change-Id: I09ab0de3094940cf804f917a0c26d0f4f56ebff7