Setup prometheus with OLM for operand monitoring
Add makefile entry to deploy the prometheus operator.
Make prometheus-operator a requirement and install it via sfconfig CLI.

Important: note that installing the prometheus operator via OLM is not supported
in MicroShift. In order to do so one has to grant privileges to the
prometheus-operator service account. This step is covered in the
Makefile entry, but must be done manually if installing the sf-operator
via OLM. The easiest way to do so is to run the new CLI command below.

Add CLI command (sfconfig prometheus [--fqdn XXX]) to install the
operator if needed, then start a prometheus instance in the sf namespace.

The instance is configured to monitor any ServiceMonitor, PodMonitor and
PrometheusRule created in the sf namespace with a label key named "sf-monitoring".

The deployed instance is available at https://prometheus.<fqdn>

Add simple test to ensure the prometheus object is properly created by
the CLI + the prometheus instance is reachable at the expected URL.

Change-Id: Ifccbab9f9303b293498c548aaeb0a6d870b421ff