Move managesf-resources dep as a gerrit sidecar container
This change is a step to avoid managesf-* and remove gerrit from
the controller.

Having the managesf-resources sidecar container part of the
gerrit deployment will still keep the capability for us and the CI
to create gerrit repos/acls/groups easily.

The *sfconfig gerrit* tool also benefits from that change.

Change-Id: I1f91f5dde6b035fe2d57ba2d4ec45bdcdbf57ea5