config repo - repository location settings update
No assumption is made on the config repo location scheme. The SF
operand no longer expects write access to the config repo then we
can only rely on a public / open access to the config repository.

This patch provides capability to pass config repo connection info
to the SF Spec. The config repository is not longer named 'config'
and can be set. Still by default we use the local gerrit with a
default to 'config' repository name.

Furthermore the name of the Zuul connection to use for Zuul to
connect on the config repository can be set.

Change-Id: Ia64ba246f8f16aabf3fb113b1de65fa2dafc4725