Zuul - Gerrit connection - avoid need for host alias
This change removes the need for setting a host alias to enable
the local resolve from Zuul pods of "gerrit.FQDN" to the internal
gerrit pod IP.

As we are removing Gerrit from the sf-operator control this alias
is an issue.

As we use dnsmasq on our dev plaform we have a resolv from internal
pods to *.FQDN so the alias can be simply removed.

This change also adds for the internal gerrit option to not verify
the certificat issuer.

Change-Id: Idcb1e822a426831265fbaa3a33ad9547f0b8f4cd