Ensure a SF operand can start without a config repo
This change permits a SF to be spawned w/o a config repository set.
As not specified the config repo is not added to the tenant config.

Then no config-check/config-update jobs are available.

This allows to get the SF operand deployed and simplify requirements
to "bootstrap" an operand. A logical way to finalize a setup (when config
repo is on a gerrit instance) is:

- gather the Zuul public key to setup a gerrit CI account
- setup the Zuul gerrit connection where is located the
  "config" repository
- finally tell the operand the config repo location (name, url, zuul
  connection name)

This change also allows to switch the config repository locations easily.

Change-Id: Ic4166aea23e5de547cf36b6e719e39af62a5ca6f