Add ConfigCheckJob Custom Resource
A ConfigCheckJob resource can be used to run "zuul-admin
tenant-conf-check" against a zuul tenant configuration on a
zuul-scheduler container. A TTL can be set (default to 600s)
to autodelete the CR and associated resources.

The CR's status has a PodID entry that can be used to fetch
the job's logs.

Change-Id: I3afd1208b01f102d7525e83d35487ac49f9fd935