clean-installation: ensure gerrit pvc are not wiped
When running "clean-installation" after "start-gerrit", the role
attempts to remove all pvcs in the sf namespace, including the
gerrit pvc. Termination however is blocked since the gerrit stateful set
is not removed.

Add a filter in the clean-installation task removing PVCs so that
the gerrit pvc is spared. If you need to remove it, use instead
the 'sfconfig gerrit -wipe' command.

Change-Id: Iaa9313eb6c9178f9218571257eb005f83d446fa7