nodepool-launcher: ensure service restart when providers secret changes
This patch introduces the Owns function [1] which triggers the reconcile
loop when a Owned (ControllerReference) watched resource (here secrets) are
changed.

Using this, the controller react to external change to the
nodepool-providers-secrets secret. As soon as the user update the
kube.config or clouds.yaml then the controller react by restarting
nodepool-launcher.

This changes also create a place holder for the that secret named
"nodepool-providers-secrets" to ensure the secret resource is
created with the right ControllerReference.

The following command could be used to update the kube.config key
of the secret:

```
kubectl patch secret nodepool-providers-secrets --patch="{\"data\": {
\"kube.config\": \"$(cat /home/fedora/.kube/config |base64 -w0)\" }}"
```

[1]: https://book.kubebuilder.io/reference/watching-resources/operator-managed.html

Change-Id: I58dd5dd21b7f2bdfb35fb70b832863eaad44b948