MariaDB: move zuul db & user creation to mariadb controller
This is preparatory work to enable using an externally created database
with Zuul. The database and user creation that used to be done in an
initContainer handled by the zuul controller is now done as a post-init
batch job. When the batch job has completed, create a secret holding
connection data, in the fashion of the mariadb-operator's connection
resource.

Change-Id: I19bcf3f5a1243bdcf41b28d72791fefe3274a521