Relax zuul-scheduler pod failure when wrong config location
This change ensures that the generate-tenant-config script when running via the zuul-scheduler pod's init-container won't fail when the config location is unaccurate.

Having the init-container to success even when the main.yaml file cannot be fully computed ensures that zuul-scheduler pod is in the running state.

Indeed, without that change it is impossible (except manually) to recover after a mistake in the config location setting. Three components must restart when the location change: git-server, nodepool, zuul-scheduler. While components restart work well for git-server and nodepool it is broken for zuul-scheduler.

According to https://kubernetes.io/docs/concepts/workloads/controllers/statefulset/#forced-rollback a rollout will only happen when the Pod is Ready which is not the case when the zuul-scheduler pod failed to start because of the init-container failure.

More insights here: https://github.com/kubernetes/kubernetes/issues/67250

The approach implemented in this patch is quite simple as it is just about to bypass the init-container failures and ensures the pod start. The idea is to make it easy to just recover from a wrong config location setting.

Change-Id: I6885605d30b5ffc8c4c327dc4590b56316bd7955