Add liveness and startup probes
From time to time we see in the CI logs, that services returned HTTP
error 500 or 503. It might be related to wrong setting startup, liveness
and readiness probe [1].
Base on the probes done by the Openstack K8S Operators community [2],
we should have different timing and threshold depending what probe is
used.

[1] https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-startup-probes/
[2] https://github.com/openstack-k8s-operators/nova-operator/blob/main/pkg/novaapi/deployment.go#L43

Change-Id: I2df39121d9cb8406d48782ffdbac5c4cd4899b23