Remove dependency to mariadb-operator
We used the mariadb-operator module only to compute the database
source name, so that the database secret is consistent with ones
issued by the operator should we switch to using it to deploy the
database. The module however has a dependency issue with
controller-runtime that makes managing our go.mod/go.sum files
difficult.
Remove this dependency and compute the DSN using the go-mysql module.

Change-Id: Ib34d17b2eedae8d80fc7f7cb22c956d7f85f7597