tests - Add a validation for cert-manager let's Encrypt
This change adds a test case that partialy validates the let's encrypt
setup flow offered by the operator.

The test updates the CR to enable let's Encrypt and then expects that
'Challenges' resources are set but not Ready for a specific
reason (NXDOMAIN error).

Indeed the http01 challenge cannot succeed as in the CI
zuul/nodepool/logserver.sftests.com are not resolvable from the outside.

Finaly the test, removes the let's encrypt setting from the CR and expects
the SF resource to become Ready.

Change-Id: I44fcc7839acf0a942b90e6acb21b664c13c5988a