Fix nodepool-providers-secrets volume mounts and bump to sf-operator to v0.0.5
nodepool-launcher raises errors in the logs such as:

```
IsADirectoryError: [Errno 21] Is a directory: '/var/lib/nodepool/.kube/config'
```

when the kube.config key is not present in the secret.

So this change re-enables conditional statemenents to volume mount
the secrets data.

Also as this prevents nodepool-builder to process jobs when no valide
kube.config configuration is set. This change also proposes to bump
to v0.0.5 in order to release a version of sf-operator.

Change-Id: Id8dd5f9808deb3fd3bb3162f97568505bb084464