Tests: move start-prometheus role to after starting the operator
We've experienced odd errors with the installation of the
prometheus-operator, when done "manually" (as opposite to as a
dependency of the sf-operator in OLM). Basically the installPlan
associated to the subscription is set to manual approval - when so far
approvals have always been automatic.

Let's see if letting OLM handle the installation fixes things.

Change-Id: I8b190e14cadf0afb5ff420d2cfd4cdab678f273a