Get namespace events: use a more "syslog-y" formatting
The `kubectl get events` command formats time references relatively to
the time at which the command was run (ie "XXX ago"). This is hard to
correlate with other logs that are timestamped properly. To make things
worse, the events aren't even displayed in chronological order.

Modify the output to display an absolute timestamp rather than a
relative one, and order events chronologically.

Change-Id: I104b2e102b9a6cbdc85271fe80a54797760442ab