nodepool-builder: enable config-check and config-update
This change adds the config-(check/update) workflow for Nodepool
builder.

The assumption is that the builder configuration is stored into
config/nodepool/nodepool-builder.yaml (which is sperated from
nodepool launcher).

The change adds a config workflow test for Nodepool builder.

The change does not validate the "build" capacity of the builder
yet. A follow-up change might add that validation.

Change-Id: I6ce48a2f3d9f3f4e66e22cf32559a05b84ef64cd