nodepool-builder: add dib-ansible wrapper script
This change adds the script dib-ansible and setup a dummy tests
that validate that an build via ansible-playbook is performed.

The script is similar to dib-virt-customize.py script that we
bundle with sf-3.8 expected to be renamed and adapted to tooling
paths accoding the builder image.

This changes does not provide a full image build worklow test. However
based on it a followup change might implement of more complete
test.

Change-Id: Ia3fdf96bbc4d5bdb78dfb7573c22c182219359f5