[WIP] AlertRules: fix nodepool OpenStack API alerts, improve lifecycle
Add a checksum of the alerts on PrometheusRule objects, so that
we know to update them if the rules changed.

There was an error in Nodepool's documentation that lead to OpenStack
API metrics not being collected properly. The metrics configuration
actually depends on the contents of clouds.yaml. From now on, the
statsd mapping config file is a template; we get the proper mappings and
alert rules by parsing clouds.yaml if it exists.

TODO

* Document feature and how to set clouds.yaml
* Testing

Change-Id: I24eb3962378008cc63c2cabd854f8f58dfb2ac1f