generate-config: ensure the config repo checkout in ~/config directory
The choosen config repo name is up to the administrator (like microzuul-config) in our first use case.

However to simplify we can simply assume that the checkout is done in a well known directory such as "config".

The "dib-ansible" command assume a checkout named "config".

So this change ensures the checkout is always named "config".

Change-Id: I810f85246d1229cc06383aa7ccb91d9249f79047