crd - logserver - avoid using CEL and stick to standard OpenAPI validation
This change makes use the standard OpenAPI validation facilities [1] and defaulting [3] to validate the LogServerSpec
object.

It also removes the use of x-kubernetes-validations which is still beta and is not garanty to be enabled.

The default values for retentionDays and loopDelay comes from the CRD.

The test_crd test case is also removed as we no longer use CEL validation, and because the others
not related tasks were left overs.

[1] https://kubernetes.io/docs/tasks/extend-kubernetes/custom-resources/custom-resource-definitions/#validation
[2] https://kubernetes.io/docs/tasks/extend-kubernetes/custom-resources/custom-resource-definitions/#validation-rules
[3] https://kubernetes.io/docs/tasks/extend-kubernetes/custom-resources/custom-resource-definitions/#defaulting

Change-Id: Iba7913733b9e6bde1748cc4122559b81349a7cdc