Add the standalone mode for running a SofwareFactory resource reconcile
The purpose of that mode is to provide a trade-off to deploy a testing Software Factory
without the need for cluster admin (for CRD installation), or the need for OLM
enabled cluster to install the sf-operator Operator.

This mode is designed only for developer context and easy startup with sf-operator
to test Software Factory deployment as a non 'cluster-admin' user.

For instance it is a convenient mode, to quickly validate that a regular
deployment of a Software Factory via the sf-operator (installed via OLM)
will work. This mode deploys all resources for a working SF deployment
similarly to how the sf-operator's controller does.

Note that adding this modes does not have impact on the regular operator
mode.

Change-Id: I3a27450100011c84e553beafcc5c5c7bfa040075