ci - nodepool-builder - Add task to copy Zuul SSH public key on image-builder
This change completes the documentation by showing an Anisble task
that demonstrate how to copy the Zuul public key on the
image-builder. The copy must then be included into the built image,
but this procedure is out of the scope of the doc.

Also this change ensures we can access the key from a cloud image
build playbook by adding that task inside the test playbook.

Change-Id: Ic3423605717ee5c885471f7d90e6d5360dbe8cb5