Inject dev ssh keys when a variable is set
Currently, developers that want to connect to the CI VM to verify
what is happening with the environment need to make many "hoops"
to connect to that host.
That feature will be helpful for developers of sf-operator project to
inject pub keys just by setting:

    inject_dev_ssh_keys: true

in the zuul.d job definition.

Change-Id: I234cb597d61c87abe884ecc7826419f954ed1158