mariadb - use a Service Pod Selector to always stick to mariadb-0 Pod.
This change makes use of a Service Pod Selector based on the Pod name instead
of label based selector. This ensures that even if the MariaDB statefulset is
scaled, consumers (such as Zuul) will only be able to contact the Pod-0 and
won't be round-robin to other Pod of the statefulset.

When a StatefulSet is scaled, each Pod get its own Volume Storage based
on the PersistentVolumeClaim. Also, the Service Resource in front of the
StatefulSet will round-robin each connection to Pod-N.

So better to ensure we stick to the first Pod-0 of the sts to avoid
unexpected behavior in case of scale of MariaDB Statefulset.

Also this change improves the connection details passing to Zuul.

Change-Id: I17646d84f044afb30dd86fc0428d7ebeaaa99088