logserver - define one Service instead of three
This change refactors Service Resource definition for the logserver.

Also, it updates the function that trigger the Zuul internal tenant reconfigure in
order to ensure that when the internal tenant's secret's serial value is updated then we run
the reconfigure command.

As we use the internal DNS name of the Service and as the Service has been renamed just 'logserver',
then the logserver's Secret content changed. Then during an upgrade the reconfigure must be
triggered.

Change-Id: I087ffcad1cda731d47205798e68f1b18d27d9cb9