Log Forwarding: enable forwarding nodepool build logs
If log forwarding is enabled, create a fluent bit sidecar on the
nodepool-builder statefulset that forwards logs a json payloads,
ie the way Nodepool and Zuul process do too.

Include a simple Ansible output callback plugin to timestamp
each line in ansible-playbook's output, which allows these log
lines to be queried in order in the log collector.

Change-Id: I5b4779b58f72aab41de3dd3262f0e1d8f90d5566