zuul-executor - only update podSpecTemplate to keep configured replicas count
This change fixes the too deep update of a the Executor sts and now
prevents to manually set Replicas count to be overwritten during a
update/rollout.

Also it update the zuul-components test role in order to ensure that
at the end of the test we keep 2 executor replicas up and running.
The scale-resources roles scales up the executor to 2 replicas at
the very beginning of the fonctionnal tests suite.

Change-Id: I2af6c8a89ea254e1d303be43deacbcff9222f210