CLI main.go: Add support for cli config file
A YAML config file listing contexts can be passed to the CLI. The config
file contains parameters that can be used by the CLI. Several contexts
can be defined.

CLI arguments can override a context.

Rename the "standalone" subcommand as "apply" as per ADR12.
Rename the default command to "operator" to run the operator.

The "apply" subcommand supports the config file parameters.

Change-Id: Ie4bff45ec748985c6d13ae4c3d5dadfc374b912c