Avoid controller panic in case of empty connection secret name
Now the GetSecret function returns a null string data (empty string)
when the secret name of the connection is "". Also we add some
logging message to raise the issue w/o panic.

This can happen in case the 'secrets' name is not
provided in the connection definition.

Also:
    - Rename GetConfigMapByNameRef to GetConfigMap
    - Rename GetSecretByNameRef to GetSecret

Note that requiring the 'secrets' key as mandatory would make sense
as apiToken and webHookToken are mandatory to get a working connection.

Change-Id: If0bcee52df0ed86a9c7f957569e7d95306c04394