Move namespace creation, dependent operators installation to playbook
The `sfconfig dev prepare` command is meant to be run multiple times,
however it includes tasks that need to be performed only once, like creating
namespaces and installing the cert-manager and prometheus operators.

Move these tasks into a playbook in tools/microshift and include this
playbook in the current `sfconfig microshift` command (it will be
moved to the new unified CLI later).

Change-Id: I5dd1970391dcab1837359f4c65f1ee9a84127425