This change removes the need of cert-manager for zookeeper
This change removes the certificates infrastructure generated by cert-manager.
Replace it with certificates generated by the `crypto` library.

This is the first step to remove a requirement that the cluster runs the
cert-manager operator, as now cert-manager's resources such as 'Certificate' are now longer
needed.

This changes updates the serial annotation of all StatefulSet and
Deployment that mount the zookeper-client-tls and
zookeeper-server-tls (Zuul, Nodepool, Zookeeper, Weeder) so that
certificates are updated on all services if needed. This means that
these services will be rerolled and pods will be recreated at upgrade
time.

Change-Id: Iace0b3aba036bc20b6db0e7c9755ee8f902dffb7