Add feature: restore
This commit provides a feature: restore - which will be helpful to
restore data from a backup.
Some part of restoring solution require kubectl binary to be installed
and configured .kube/config file as well.
With that way, we don't need to copy the restore file/dir to the pod or
create new pod with new PV or mount same PVC into new pod, which might be
rejected by some PV drivers. Also mounting local host directory
to the OpenShift cluster might be prohibited in some
deployments (especially in public deployments where user is not an admin), so
that is not a good idea to use.

Change-Id: I7c44bb18dcbd55895df30d9ee5add1d6c42a7625