Fluent Bit log forwarding: fix erroneous labels, add container label
Add a "container" label that makes it easier to harmonize labels between
Promtail collection or integrated Fluent Bit log forwarding.

If Promtail is running as a service on nodes, the pods' logs are parsed
in a way that exposes the following labels: namespace, pod, pod_uid,
container.
While the "pod uid" info would probably require a query to the
kubernetes API to figure out, this change ensures that at least the
namespace, pod and container labels will be set consistently when using
the fluent bit forwarder, so that label queries can be written and work
regardless of how logs are collected.

Change-Id: Ib5f6def263bd9c1b1fd6f95001746afad43192fc