Remove Route handling
As Route/Ingress are dependent of the cluster and because how TLS might
be configured and handled it is safer to keep that burden outside of
the sf-operator control.

This change acts in that direction by removing the Route/TLS/Let'sEncrypt
support.

The sf-operator dev CLI, `create demo-env` ensure the Route resource
exists.

The doc is amended to reflect that change, and a new ADR is added.

Change-Id: Ia71077b028223a878206c37a0a6cd5fddac885d1