Update MariaDB container image to 10.6
We can not perform upgrade directly from 10.5 to 11, but we need
to make a special steps to perform an upgrade.
According to the docs [1][2], we need to make an upgrade to 10.6 first,
then we can try to go to MariaDB 11.4, but we should consider to upgrade
to 10.11 first, then go to the 11.4.

[1] https://mariadb.com/kb/en/upgrading/
[2] https://mariadb.com/kb/en/upgrading-from-mariadb-10-11-to-mariadb-11-4/

Depends-On: https://softwarefactory-project.io/r/c/containers/+/31604

Change-Id: I413498c0afad3a37eea69dc5723e288579e54482