Switch MariaDB log from general log to error log
The general_log is providing a lot logs, that later are breaking logcli
tool. We don't need to grab all logs, but just more important which are
error_log messages.
The fluentbit forwarder sidecard container for MariaDB is monitoring
only logs with extension: '.log', so let's create such file, so the
health check tests whould be also satisfied with that change.

Change-Id: Ie3e4a8c9bcff12c86fed5c5acfa14d32e9b77245