Bump MicroShift default version to 4.14; increase Gerrit memory limit
This also require to update olm and operator sdk to newer release.
After making few tests, on making the health-check tests, the
Gerrit container is killed by the OOMKill, which raise an issue, due the
Zuul did not read the stream, so the test is waiting to run the Zuul
job, but it is not triggered because of the died Gerrit.

NOTE: The zuul-ext test needs to change the k8s-api-url to the internal
URL, due the https://{{ fqdn }}:6443 stops working with the CA cert
provided by the serviceaccount (mounted in /run/secrets/kubernetes.io/serviceaccount/ca.crt
when the automountServiceAccountToken is set to true).
The error is correct, because in the MicroShift 4.13 and earlier the
cert contains info also for public Kubernetes API url.

Change-Id: I79cc2b7d8a4adb3ef0594ed5766fadb2cd064767