Zuul/Nodepool/ZK - extend internal certficiate validity duration
We are issuing those internal certificates using cert-manager for
the ZK <-> Zuul/Nodepool SSL communication.

The default validity for them was 3 months with a renewal after 2 months.
The issue is that we need to handle the restart of Nodepool/Zuul and
ZK when the renewal is done by cert-manager.

This patch proposes to avoid that burden by extending the validity
to a very long duration.

If for whatever reason the local CA and belonging
certificates need to be rotated, checkout the sf-operator documentation
to read about the process.

Change-Id: I1bc1885589e11f8300a2d46fe9e53737c5407398