restore - ensure shutdown of logserver after restore
The logserver-ssd container part of the logserver pod requires the
"zuul-ssh-key" secret to initialize the authorized key capable of
pushing on the logserver.

After a restore the logserver-sshd must be restarted to update the
key passed as an ENV var. Thus the restore command must stop the
logserver to force the reconcile loop of the operator to re-create
the pod definition and start it.

Change-Id: I376d1fc3fd96317fd5800de0fa79cd0406275d3e