Increase Zuul Scheduler and Zuul Web Startup Probes Time
This change increases the values of the startup probes on Zuul
Scheduler and Zuul Web, this way it allows Zuul Scheduler long
repositories loads, and Zuul Web to wait for Zuul Scheduler to finish.
It sets a total time of 3600s ( 1 hour ) at startup.
Periond of Seconds for each Probe Check
Failure Threshold for the Probe Check allow failures before
terminating and restarting the Pod.

Sources:
https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-startup-probes/
https://kubernetes.io/docs/concepts/configuration/liveness-readiness-startup-probes/
Change-Id: I0a405c5443f0705ea7cb5f7d681d6295dc817ae3