Add script to simplify the upgrade of Zuul version
Upgrading sf-operator to use a new Zuul release involves manual
updates to multiple repositories, requiring time-consuming checks of
Git logs to determine which files need changes and how to modify them.

This script reduces the complexity of the process by providing a
structured way to update hashes and versions in the sf-operator and
containers repositories, saving time and minimizing the risk of errors
during the upgrade.

- Updates hashes and versions in sf-operator
and containers repositories.

Change-Id: I7152bf392998391fb101f619c46b8f1b5ae49d01