hound - remove the CODESEARCH_ENABLED env var
This change removes the env var toogle that define if codesearch is
enable (w/r to the config-update playbook) as this does not work out
of the box when the service is disabled. The git-server is not
restarted with the new env var.

This change simplifies the playbook by simply checking for the pod
and decides to trigger the pod deletion (to force sts to handle the
restart) only if the pod exists. This way we don't need such a
toogle var.

Also, this change removes the in-place config refresh of the hound
config because in any cases restarting the pod the startup regenerates
the config from the tip of the config repository.

Finally, do not rely on the post-tasks as there is not reason for this.

Change-Id: I474630547c7fa377e418d90bd41155c819cfee68