Removal of cert-manager dependency, part 2 of 3
By v0.0.55 we got rid of resources created with cert-manager.
This change removes the code dependency, fixes outdated documentation,
and remove the installation of the cert-manager operator on test/dev
envs.

Since the "olm-deps/cert-manager.yaml" file is still used when
deploying v0.0.55, this file will be removed after the v0.0.56 tag.

Change-Id: I07391452f0b0fe92376b0e0f83afdd0c93b037cd