Flakyness: backup/restore - The 'internal' tenant contains error(s) or warning(s)
```
Zuul encountered a syntax error while parsing its\nconfiguration in the repo system-config on branch master.  The\nproblem was:\n\n  Decryption failed\n\nThe problem appears in the the \"base\" job stanza:
```

This change attempts to delete the zk state after the restored
happened and before the restart of the component.

This command is supposed to be run when zuul services are stopped.

Change-Id: I1485848a4e64d01ee7b4cb90825fdfd77cc4a335