Replace k8s_info with kubectl command to fetch nodepool-launcher
This change fixes the following config-update error:

  fatal: [localhost]: FAILED! => {"changed": false, "msg": "Exception 'HTTPSConnectionPool(host='cluster', port=6443): Max retries exceeded with url: /api/v1/namespaces/sf/pods?labelSelector=run+%3D+nodepool-launcher (Caused by SSLError(SSLCertVerificationError(1, '[SSL: CERTIFICATE_VERIFY_FAILED] certificate verify failed: unable to get issuer certificate (_ssl.c:1028)')))' raised while trying to get resource using {'namespace': 'sf', 'label_selector': 'run = nodepool-launcher'}"}

… which seems to happen when the in-cluster kubeconfig certificate-authority doesn't match
the public API. In that case, using the kubectl command works, but the Ansible Python module
is not able to verify the TLS certificate. It appears that the Python implementation doesn't
use the same logic as the golang client, for example it also requires the namespace to be
provided, even though it is set in the kube config current context.

Therefor, this change replaces the Ansible kubernetes module with a simpler kubectl command
to avoid discrepancies between the clusters.

FIXME: provide a reproducer and report the issue upstream
Change-Id: I1eadd45ee9293b5fb932de0f1b8abc5fc3c2e0ad